search function takes search text as a new document remake tfidf vectors.
