+++ code/document_db.py
@@ -4,17 +4,16 @@ import create
 import search
 
 
-def invalido():
-    print("invalid command")
-    exit()
+def invalid():
+    quit("invalid command")
 
 
 if len(sys.argv) != 3:
-    invalido()
+    invalid()
 
 if sys.argv[1] == "-create":
     create.create(sys.argv[2])
 elif sys.argv == "-search":
     search.search(sys.argv[2])
 else:
-    invalido()
+    invalid()
